% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightgbm.R
\docType{data}
\name{agaricus.train}
\alias{agaricus.train}
\title{Training part from Mushroom Data Set}
\format{
A list containing a label vector, and a dgCMatrix object with 6513
rows and 127 variables
}
\usage{
data(agaricus.train)
}
\description{
This data set is originally from the Mushroom data set,
             UCI Machine Learning Repository.
             This data set includes the following fields:

              \itemize{
                  \item{\code{label}: the label for each record}
                  \item{\code{data}: a sparse Matrix of \code{dgCMatrix} class, with 126 columns.}
               }
}
\references{
https://archive.ics.uci.edu/dataset/73/mushroom

Bache, K. & Lichman, M. (2013). UCI Machine Learning Repository
[https://archive.ics.uci.edu/]. Irvine, CA: University of California,
School of Information and Computer Science.
}
\keyword{datasets}
