% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgb.Dataset.R
\name{colnames.lgb.Dataset}
\alias{colnames.lgb.Dataset}
\title{Handling of column names of \code{lgb.Dataset}}
\usage{
colnames.lgb.Dataset(x)
}
\arguments{
\item{x}{object of class \code{lgb.Dataset}}
}
\description{
Handling of column names of \code{lgb.Dataset}
}

