% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgb.Dataset.R
\name{colnames.lgb.Dataset}
\alias{colnames.lgb.Dataset}
\title{Get column names of \code{lgb.Dataset}}
\usage{
colnames.lgb.Dataset(dataset, ...)
}
\arguments{
\item{dataset}{object of class \code{lgb.Dataset}}

\item{...}{other parameters}
}
\description{
Get column names of \code{lgb.Dataset}
}

