% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgb.Dataset.R
\name{construct.lgb.Dataset}
\alias{construct.lgb.Dataset}
\title{Construct Dataset explicit}
\usage{
construct.lgb.Dataset(dataset)
}
\arguments{
\item{dataset}{Object of class \code{lgb.Dataset}}
}
\description{
Construct Dataset explicit
}
\examples{
data(agaricus.train, package='lightgbm')
train <- agaricus.train
dtrain <- lgb.Dataset(train$data, label=train$label)
construct.lgb.Dataset(dtrain)
}

