% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgb.Dataset.R
\name{create.valid.lgb.Dataset}
\alias{create.valid.lgb.Dataset}
\title{Contruct a validation data according to training data}
\usage{
create.valid.lgb.Dataset(dataset, data, info = list(), ...)
}
\arguments{
\item{dataset}{lgb.Dataset object, training data}

\item{data}{a \code{matrix} object, a \code{dgCMatrix} object or a character representing a filename}

\item{info}{a list of information of the lgb.Dataset object}

\item{...}{other information to pass to \code{info}.}
}
\description{
Contruct a validation data according to training data
}
\examples{
data(agaricus.train, package='lightgbm')
train <- agaricus.train
dtrain <- lgb.Dataset(train$data, label=train$label)
lgb.Dataset.save(dtrain, 'lgb.Dataset.data')
dtrain <- lgb.Dataset('lgb.Dataset.data')
}

