% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgb.Dataset.R
\docType{data}
\name{Dataset}
\alias{Dataset}
\title{Dataset class of LightGBM}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Dataset
}
\description{
The Dataset used in LightGBM is a high memory efficient object.
The LightGBM only accept Dataset object for training
}
\keyword{datasets}

