% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgb.convert.R
\name{lgb.convert}
\alias{lgb.convert}
\title{Data preparator for LightGBM datasets (integer)}
\usage{
lgb.convert(data)
}
\arguments{
\item{data}{A data.frame or \code{data.table} to prepare. If a \code{data.table} is provided,
it will be modified in place for speed and too avoid out-of-memory erros.}
}
\value{
The cleaned dataset. It must be converted to a matrix format (\code{as.matrix})
        for input in \code{lgb.Dataset}. If \code{data} is a \code{data.table}, it will be
        modified in place.
}
\description{
Attempts to prepare a clean dataset to put in a \code{lgb.Dataset}.
             Factor, character, and logical columns are converted to integer.

             Missing values in factor and character columns will be replaced with 0. Missing
             values in logical columns will be replaced with -1.

             Please use \code{\link{lgb.convert_with_rules}} if you want to apply this
             transformation to other datasets.

             NOTE: In previous releases of LightGBM, this function was called \code{lgb.prepare}.
}
\examples{
data(iris)

str(iris)

# Convert all factors/chars to integer
str(lgb.convert(data = iris))

\dontrun{
# When lightgbm package is installed, and you do not want to load it
# You can still use the function!
lgb.unloader()
str(lightgbm::lgb.convert(data = iris))
# 'data.frame':	150 obs. of  5 variables:
# $ Sepal.Length: num  5.1 4.9 4.7 4.6 5 5.4 4.6 5 4.4 4.9 ...
# $ Sepal.Width : num  3.5 3 3.2 3.1 3.6 3.9 3.4 3.4 2.9 3.1 ...
# $ Petal.Length: num  1.4 1.4 1.3 1.5 1.4 1.7 1.4 1.5 1.4 1.5 ...
# $ Petal.Width : num  0.2 0.2 0.2 0.2 0.2 0.4 0.3 0.2 0.2 0.1 ...
# $ Species     : int  1 1 1 1 1 1 1 1 1 1 ...
}

}
