% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgb.Booster.R
\name{lgb.dump}
\alias{lgb.dump}
\title{Dump LightGBM model to json}
\usage{
lgb.dump(booster, num_iteration = NULL)
}
\arguments{
\item{booster}{Object of class \code{lgb.Booster}}

\item{num_iteration}{number of iteration want to predict with, <= 0 means use best iteration}
}
\value{
json format of model
}
\description{
Dump LightGBM model to json
}

