% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgb.interpret.R
\name{lgb.interprete}
\alias{lgb.interprete}
\title{DEPRECATED - use lgb.interpret() instead}
\usage{
lgb.interprete(...)
}
\arguments{
\item{...}{Arguments passed through to \code{lgb.interpret}}
}
\description{
Alias for \code{lgb.interpret}.
}
