% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lgb.params2str}
\alias{lgb.params2str}
\title{convert params to str that can be accepted by lightgbm c_api}
\usage{
lgb.params2str(params, ...)
}
\description{
convert params to str that can be accepted by lightgbm c_api
}

