% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removed.R
\name{lgb.prepare}
\alias{lgb.prepare}
\title{removed functions}
\usage{
lgb.prepare(...)
}
\arguments{
\item{...}{catch-all too match old calls}
}
\value{
Nothing. This function always raises an exception
}
\description{
removed functions
}
