% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgb.Dataset.R
\name{lgb_shared_dataset_params}
\alias{lgb_shared_dataset_params}
\title{Shared Dataset parameter docs}
\arguments{
\item{label}{vector of labels to use as the target variable}

\item{weight}{numeric vector of sample weights}

\item{init_score}{initial score is the base prediction lightgbm will boost from}

\item{group}{used for learning-to-rank tasks. An integer vector describing how to
group rows together as ordered results from the same set of candidate results
to be ranked. For example, if you have a 100-document dataset with
\code{group = c(10, 20, 40, 10, 10, 10)}, that means that you have 6 groups,
where the first 10 records are in the first group, records 11-30 are in the
second group, etc.}
}
\description{
Parameter docs for fields used in \code{lgb.Dataset} construction
}
\keyword{internal}
