% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgb.Predictor.R
\docType{data}
\name{Predictor}
\alias{Predictor}
\title{Predictor of LightGBM
Not expose, only for package internal use}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Predictor
}
\description{
Predictor of LightGBM
Not expose, only for package internal use
}
\keyword{datasets}

