% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgb.Booster.R
\name{print.lgb.Booster}
\alias{print.lgb.Booster}
\title{Print method for LightGBM model}
\usage{
\method{print}{lgb.Booster}(x, ...)
}
\arguments{
\item{x}{Object of class \code{lgb.Booster}}

\item{...}{Not used}
}
\value{
The same input \code{x}, returned as invisible.
}
\description{
Show summary information about a LightGBM model object (same as \code{summary}).

             \emph{New in version 4.0.0}
}
