% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgb.Dataset.R
\name{save.binary.lgb.Dataset}
\alias{save.binary.lgb.Dataset}
\title{save \code{lgb.Dataset} to binary file}
\usage{
save.binary.lgb.Dataset(dataset, fname)
}
\arguments{
\item{dataset}{object of class \code{lgb.Dataset}}

\item{fname}{object filename of output file}
}
\description{
save \code{lgb.Dataset} to binary file
}

