% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgb.Dataset.R
\name{set.categorical.feature.lgb.Dataset}
\alias{set.categorical.feature.lgb.Dataset}
\title{set categorical feature of \code{lgb.Dataset}}
\usage{
set.categorical.feature.lgb.Dataset(x, categorical_feature)
}
\arguments{
\item{x}{object of class \code{lgb.Dataset}}

\item{categorical_feature}{categorical features}
}
\description{
set categorical feature of \code{lgb.Dataset}
}

