% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgb.train.R
\name{lgb.train}
\alias{lgb.train}
\title{Main training logic}
\usage{
lgb.train(params = list(), data, nrounds, valids = list(), obj = NULL,
  eval = NULL, verbose = 1, eval_freq = 1, init_model = NULL,
  colnames = NULL, categorical_feature = NULL,
  early_stopping_rounds = NULL, callbacks = list(), ...)
}
\description{
Main training logic
}

