% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightgbm.R
\name{lightgbm}
\alias{lightgbm}
\title{Simple interface for training an lightgbm model.
Its documentation is combined with lgb.train.}
\usage{
lightgbm(data, label = NULL, weight = NULL, params = list(),
  nrounds = 10, verbose = 1, eval_freq = 1L,
  early_stopping_rounds = NULL, save_name = "lightgbm.model",
  init_model = NULL, callbacks = list(), ...)
}
\description{
Simple interface for training an lightgbm model.
Its documentation is combined with lgb.train.
}
