% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgb.Dataset.R
\name{set.reference.lgb.Dataset}
\alias{set.reference.lgb.Dataset}
\title{set reference of \code{lgb.Dataset}}
\usage{
set.reference.lgb.Dataset(x, reference)
}
\arguments{
\item{x}{object of class \code{lgb.Dataset}}

\item{reference}{object of class \code{lgb.Dataset}}
}
\description{
set reference of \code{lgb.Dataset}
}

